Added a test for getting the zone when sending a string.
Who knows the longitude and latitude of where they are without
having to look it up?

In addition to accepting a lat and lng this api should also
accept a string which will be the underground station
a user is near or knows they are near or wants to know
what zone a station is in.