Tests pass
Changed route in test_get_zone_by_station name on the basis that a
string can be submitted to get the zone.

A nice-to-have will be that the long and lat are also returned with
the zone.

Not yet worked out how to test that param submitted is passed to view.